Support Strata Source expanded limits
Strata Source (BSP version 25) increases many object count limits,
which necessitated upgrading many lump fields with larger data types
(many short to int upgrades). It also increases the maximum map extents
to 65536 units from the origin.

This commit was re-formed from the original found here:
https://github.com/StrataSource/bspsrc/commit/ecc46a2a8ca2c4e81fb281241c937e290402a2b7

+++ bspsrc-lib/src/main/java/info/ata4/bspsrc/lib/struct/DAreaportal.java
@@ -22,10 +22,10 @@ import java.io.IOException;
  */
 public class DAreaportal implements DStruct {
 
-    public short portalKey;
-    public short otherportal;
-    public short firstClipPortalVert;
-    public short clipPortalVerts;
+    public int portalKey;
+    public int otherportal;
+    public int firstClipPortalVert;
+    public int clipPortalVerts;
     public int planenum;
 
     @Override
@@ -44,10 +44,10 @@ public class DAreaportal implements DStruct {
 
     @Override
     public void write(DataWriter out) throws IOException {
-        out.writeShort(portalKey);
-        out.writeShort(otherportal);
-        out.writeShort(firstClipPortalVert);
-        out.writeShort(clipPortalVerts);
+        out.writeShort((short)portalKey);
+        out.writeShort((short)otherportal);
+        out.writeShort((short)firstClipPortalVert);
+        out.writeShort((short)clipPortalVerts);
         out.writeInt(planenum);
     }
 }
